requirements.txt
beautifulsoup4==4.4.1
cssselect==0.9.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.10.1
idna==2.1
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
newspaper3k==0.1.7
Pillow==3.1.1
python-dateutil==2.5.1
pytz==2016.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
tldextract==1.7.5
Werkzeug==0.11.4
